WIP, irrotettu SuoritusArvosanojen parsiminen Koskiarvosanahandlerista
